==> src/_OLD.rmd/14-merge-report.Rmd <==
# QA for merging {-}

## Purpose {-}

This page summarizes which screening/demographics data and which home visit data have anomalies or missing data.

## Screening/demographic data {-}

```{r}
screen_df <-
  readr::read_csv(
    file.path(
      here::here(),
      "data",
      "csv",
      "screening",
      "agg",
      "PLAY-screening-datab-latest.csv"
    ),
    col_types = readr::cols(.default = 'c'),
    show_col_types = FALSE
  )
```

The following rows have incomplete or missing `site_id` values:

```{r}
screen_df |>
  dplyr::filter(is.na(site_id) | is.null(site_id)) |>
  dplyr::select(site_id, vol_id, participant_ID, session_id, play_id, group_name) |>
  dplyr::arrange(vol_id, site_id, participant_ID) |>
  knitr::kable() |>
  kableExtra::kable_classic()
```

::: {.rmdnote}

Volume 1103 is OHIOS. Volume 1482 is CSUFL. Volume 954 is GEORG.
The missing values for `vol_id` indicate that there is a bug in the [cleaning code](screening-call.html).

**2023-10-20**

On closer investigation, the screening data do _not_ show an OHIOS session with `participant_ID` == '001'. There are three with '000' and two with '002'.

Similarly, for CSUFL, there is no '003' or '006'. We have home visit data for '006'.

Similarly, for GEORG, there is a '???', but no '001'.

:::

The following rows have incomplete or missing `participant_ID` values:

```{r}
screen_df |>
  dplyr::filter(is.na(participant_ID) | is.null(participant_ID)) |>
  dplyr::select(site_id, vol_id, participant_ID, session_id, play_id, group_name) |>
  dplyr::arrange(vol_id, site_id, participant_ID) |>
  knitr::kable() |>
  kableExtra::kable_classic()
```

The following rows have incomplete or missing `play_id` values:

```{r}
screen_df |>
  dplyr::filter(is.na(play_id) | is.null(play_id)) |>
  dplyr::select(site_id, vol_id, participant_ID, session_id, play_id, group_name) |>
  dplyr::arrange(vol_id, site_id, participant_ID) |>
  knitr::kable() |>
  kableExtra::kable_classic()
```

::: {.rmdnote}

There are duplicate entries for OHIOS 1103 56674 002.

We should add duplicate checking to the cleaning code.

:::

::: {.rmdnote}

There are `r stringr::str_detect(names(screen_df), "exclusion") |> sum()` variables with information about exclusion status:

```{r}
ex_dups <- stringr::str_detect(names(screen_df), "exclusion")
names(screen_df)[ex_dups]
```

These probably result from some bug in the `*_join` operation in the [cleaning process](screening-call.html#merge).
They should be merged.

:::

## Home visit data {-}

```{r}
targets::tar_load(home_visit_df, store=file.path(here::here(), "_targets"))
```

The `home_visit_df` data have some field names that are inconsistent with the screening/demographic data files.
We reconcile these differences first.

```{r}
home_df <- home_visit_df |>
    dplyr::rename("play_id" = "participant_id") |>
    dplyr::rename("participant_ID" = "subject_number")
```

Since the `home_visit_df` data have not yet been merged with Databrary information, the `vol_id` and `group_name` variables are not available.

The following rows have incomplete or missing `site_id` values:

```{r}
home_df |>
  dplyr::filter(is.na(site_id) | is.null(site_id)) |>
  dplyr::select(site_id, participant_ID, play_id) |>
  dplyr::arrange(site_id, participant_ID) |>
  knitr::kable() |>
  kableExtra::kable_classic()
```

The following rows have incomplete or missing `participant_ID` values:

```{r}
home_df |>
  dplyr::filter(is.na(participant_ID) | is.null(participant_ID)) |>
  dplyr::select(site_id, participant_ID, play_id) |>
  dplyr::arrange(site_id, participant_ID) |>
  knitr::kable() |>
  kableExtra::kable_classic()
```

The following rows have incomplete or missing `play_id` values:

```{r}
home_df |>
  dplyr::filter(is.na(play_id) | is.null(play_id)) |>
  dplyr::select(site_id, participant_ID, play_id) |>
  dplyr::arrange(site_id, participant_ID) |>
  knitr::kable() |>
  kableExtra::kable_classic()
```
