==> src/_OLD.rmd/18-post-visit-notes.Rmd <==
# Post-visit notes {-}

## Select forms from KBT {-}

The post-visit notes are stored in a separate KBT form with "Post-Visit" in the title.

```
tar_target(kb_post_visit, dplyr::filter(
    kb_df, stringr::str_detect(title, "Post\\-Visit")
  ))
```
