==> src/_OLD.rmd/24-viz-locomotion.Rmd <==
# Locomotion visualizations {-}

```{r}
if (!('home_visit_df' %in% ls())) {
  targets::tar_load(home_visit_df, store="../_targets")
}
```

## Select & Summarize {-}

```{r make-loco-df}
play_loco <- home_visit_df %>%
  dplyr::select(
    .,
    age_group,
    child_sex,
    language_child,
    site_id,
    subject_number,
    locomotor_milestones.who_walk.who_walk_onset_mo,
    locomotor_milestones.k_walk.k_walk_onset_mo,
    locomotor_milestones.crawl_onset.crawl_onset_mo
  ) %>%
  dplyr::rename(
    .,
    walk_mos_who = locomotor_milestones.who_walk.who_walk_onset_mo,
    walk_mos_kea = locomotor_milestones.k_walk.k_walk_onset_mo,
    crawl_mos = locomotor_milestones.crawl_onset.crawl_onset_mo
  ) %>%
  dplyr::mutate(
    .,
    walk_mos_who = as.numeric(walk_mos_who),
    walk_mos_kea = as.numeric(walk_mos_kea),
    crawl_mos = as.numeric(crawl_mos)
  )
```

```{r}
xtabs(formula = ~ child_sex + age_group, data = play_loco)
```

### Check for anomalous values {-}

```{r}
crawl_mos_min <- 4
walk_mos_min <- 6
```

#### Anomalous crawling onset {-}

```{r}
play_loco %>%
  dplyr::select(., site_id, subject_number, crawl_mos) %>%
  dplyr::filter(., crawl_mos < crawl_mos_min) %>%
  knitr::kable(format = 'html') 
```

#### Anomalous walking onset (KEA criteria) {-}

```{r}
play_loco %>%
  dplyr::select(., site_id, subject_number, walk_mos_kea) %>%
  dplyr::filter(., walk_mos_kea < walk_mos_min) %>%
  knitr::kable(format = 'html') 
```

##### Anomalous walking onset (WHO criteria) {-}

```{r}
play_loco %>%
  dplyr::select(., site_id, subject_number, walk_mos_who) %>%
  dplyr::filter(., walk_mos_who < walk_mos_min) %>%
  knitr::kable(format = 'html') 
```

### Crawl onset {-}

```{r fig-crawl-onset-hist, fig.cap="Age of crawling onset (mos) by sex"}
play_loco %>%
  dplyr::filter(., crawl_mos > crawl_mos_min, !is.na(crawl_mos)) %>%
  ggplot(.) +
  aes(crawl_mos, fill = child_sex) +
  geom_histogram(bins = 12) +
  theme(legend.position = "bottom") +
  theme(legend.title = element_blank())
```

### Walk onset {-}

```{r fig-walk-mos-kea, fig.cap="Age (mos) of walking onset (KEA criteria) by sex"}
play_loco %>%
  dplyr::filter(., walk_mos_kea > walk_mos_min, !is.na(walk_mos_kea)) %>%
  ggplot(.) +
  aes(walk_mos_kea, fill = child_sex) +
  theme(legend.position="bottom") +
  geom_histogram(bins = 10)
```

```{r fig-walk-mos-who, fig.cap="Age (mos) of walking onset (WHO criteria) by sex"}
play_loco %>%
  dplyr::filter(., walk_mos_who > walk_mos_min, !is.na(walk_mos_who)) %>%
  ggplot(.) +
  aes(walk_mos_who, fill = child_sex) +
  geom_histogram(bins=12) +
  theme(legend.position="bottom") +
  theme(legend.title = element_blank())
```

```{r fig-walk-mos-kea-who, fig.cap="Walking onset by WHO vs. KEA criteria"}
play_loco %>%
  dplyr::filter(., walk_mos_who > walk_mos_min, !is.na(walk_mos_who), 
                walk_mos_kea > walk_mos_min, !is.na(walk_mos_kea)) %>%
  ggplot(.) +
  aes(walk_mos_who, walk_mos_kea, color = child_sex) +
  geom_point() +
  geom_smooth(method = "lm") +
  xlim(8, 18) +
  ylim(8, 18) +
  theme(legend.position = "bottom") +
  theme(aspect.ratio = 1) +
  theme(legend.title = element_blank()) -> walk_p

ggExtra::ggMarginal(
  walk_p,
  play_loco,
  walk_mos_who,
  walk_mos_kea,
  type = "density",
  margins = "both",
  groupColour = TRUE,
  groupFill = TRUE
)
```

```{r fig-walk-mos-kea-crawl-mos, fig.cap="Walking onset vs. Crawling"}
play_loco %>%
  dplyr::filter(., crawl_mos > crawl_mos_min, !is.na(crawl_mos), 
                walk_mos_kea > walk_mos_min, !is.na(walk_mos_kea)) %>%
  ggplot(.) +
  aes(crawl_mos, walk_mos_kea, color = child_sex) +
  geom_point() +
  geom_smooth(method = "lm") +
  theme(legend.position = "bottom") +
  theme(aspect.ratio = 1) +
  theme(legend.title = element_blank()) -> walk_p

ggExtra::ggMarginal(
  walk_p,
  play_loco,
  walk_mos_who,
  walk_mos_kea,
  type = "density",
  margins = "both",
  groupColour = TRUE,
  groupFill = TRUE
)
```
