==> src/_OLD.rmd/29-post-visit.Rmd <==
# Post-visit data {-}

We load the post-visit survey data.

```{r load-post-visit-df}
tar_load(post_visit_df, store="../_targets")

dim(post_visit_df)
```

::: {.rmdnote}

Cleaning this data is set aside for future work.

:::
